+++ minidump/minidump_thread_id_map.cc
@@ -36,7 +36,7 @@ void BuildMinidumpThreadIDMap(
   for (const ThreadSnapshot* thread_snapshot : thread_snapshots) {
     uint64_t thread_id_64 = thread_snapshot->ThreadID();
     if (thread_id_map->find(thread_id_64) == thread_id_map->end()) {
-      uint32_t thread_id_32 = thread_id_64;
+      uint32_t thread_id_32 = static_cast<uint32_t>(thread_id_64);
       if (thread_ids_32.find(thread_id_32) != thread_ids_32.end()) {
         collision = true;
         break;
